my approach
I think it should end the loop once the first element can define the result rather than using priority (no need to sort on the second prop). However it should work in theory (I do not know why it fails). Plus that approach seems like a magic to me and I found it might be hard to track.